Change `AdapterInfo::{device,vendor}` fields to `u32` (#3760)
* Change AdapterInfo device and vendor to u32s

* Add CHANGELOG

+++ wgpu-hal/src/gles/adapter.rs
@@ -170,7 +170,7 @@ impl super::Adapter {
 
         wgt::AdapterInfo {
             name: renderer_orig,
-            vendor: vendor_id as usize,
+            vendor: vendor_id,
             device: 0,
             device_type: inferred_device_type,
             driver: String::new(),
